Create README.md
The entire world is transforming quickly under the present innovations. The Internet 
has become a basic requirement for everybody with the Web being utilized in every 
field. With the rapid increase in social network applications, people are using these 
platforms to voice their opinions with regard to daily issues. Gathering and analyzing 
peoples’ reactions toward buying a product, public services, and so on are vital. 
Sentiment analysis (or opinion mining) is a common dialogue preparation task that 
aims to discover the sentiments behind opinions in texts on varying subjects. In 
recent years, researchers in the field of sentiment analysis have been concerned with 
analyzing opinions on different topics such as movies, commercial products, and 
daily societal issues. Twitter is an enormously popular microblog on which clients 
may voice their opinions. Opinion investigation of Twitter data is a field that has 
been given much attention over the last decade and involves dissecting “tweets” 
(comments) and the content of these expressions. As such, this paper explores the 
sentiment analysis applied to Twitter data,